Squashed commit of the following:
commit de630beac64449d446bfed092a3ec957cdc3d3e6
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Mon Mar 2 21:50:42 2026 -0500

    Fix new warnings for missing brackets.

commit 047bba6bc4a018c2df77b8066c26edf7a3ec2e4b
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Mon Mar 2 21:34:42 2026 -0500

    Merge from upstream

commit 16080612e9c761ee3d9bcd278f3b1c8ddd1fc5ed
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Mon Mar 2 21:12:40 2026 -0500

    Merge remote-tracking branch 'upstream/master' into warnings_fixes

commit 314da46bbfda16a21e19a12e02c067d9577a26b1
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 21:21:47 2026 -0500

    Fix up basically the rest of the non-libultra warnings

commit d4ce2ed49aa6399f349af3dffcf8b13a1f2ea3e8
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 20:40:42 2026 -0500

    5 more files

commit 564474d9f3a2743472d911673fd29c58907af34d
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 20:25:05 2026 -0500

    Many warning fixes

commit 9416be34acbd9d46da57207aa361bfbae313d0ba
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 11:07:27 2026 -0500

    more warnings

commit 33d7e82e7b1f02d299726d46fed816ff20a6a80a
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 10:52:19 2026 -0500

    scheduler

commit ab947d78dd04496b420789ff3ac632acdf5773a2
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 10:35:11 2026 -0500

    Made pragma ignores a macro

commit b893e6270280f0cb1fd2dc495496da8679cd45bc
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 10:14:56 2026 -0500

    more warnings

commit 669bdbaadccb4f322e135a111f83a4945e3ec224
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 10:14:45 2026 -0500

    migrate rodata strings for font.c with stubbed printfs

commit 63c5fec2984c7b2a24d30e5bde1a54af23773ef8
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 08:28:13 2026 -0500

    queue

commit 5a7700d6270c8804e5f4e9d0eca8d385775d9378
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 08:03:53 2026 -0500

    small voxmap documenting

commit 0c1acf207bc7d0219ff09aa65da4520f0c9e9cec
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Fri Feb 27 07:55:10 2026 -0500

    voxmap warnings

commit 1d7b44cfcc231e3644e2d975e2dc41f8d4f3afa6
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 21:23:42 2026 -0500

    type update warnings

commit 4848efb406ba229eb55f237ca37036be58abaa87
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 21:09:37 2026 -0500

    map and objhits

commit 97521b747e0a5a53dd964a51765f5c423892a095
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:36:55 2026 -0500

    CC files

commit 3361f2225a6f04fbf8a02ccf22970035f53d9c80
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:24:46 2026 -0500

    bug warning

commit 359f27d8652aa7428c638391e0dea6a9e03b88b4
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:20:53 2026 -0500

    animator

commit 61a406938f7b57112981bc44eb01380a82a1c1f9
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:18:36 2026 -0500

    snowhorn

commit 63a59d343e1499af237ecd7e257bbb925ba6a7ee
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:15:03 2026 -0500

    SHmushroom

commit 397e2efa69750e770d1a486f8a558970875e0dd7
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:12:14 2026 -0500

    missing prototype

commit 35996c558ee4cec60a1b411db78c70195919ce95
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 14:09:06 2026 -0500

    Type fixes

commit d56394d8fc262290bc9521561c096eca8f3e8ec8
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 13:58:01 2026 -0500

    allow unused labels without warnings

commit 5f242187b2f21f24f0e36c16b186e338a26655c4
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 12:56:52 2026 -0500

    Fix all missing braces warnings

commit 5d70d750a84a74f424c1d638a0a5aa77de993ac4
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 11:43:50 2026 -0500

    fakematch warning fix

commit fed75b995032fa84dcbc3823624e5353ca886bbc
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 11:17:02 2026 -0500

    debris

commit 621eeedc64704fcc18cc3ce43245554c15e438d9
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 11:12:42 2026 -0500

    wmcrystal

commit fb6ec66af23a97ad161abf013ade01740c835acc
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 11:07:01 2026 -0500

    dll_703_control type fix

commit fb6b0e07d9a8d7db112967ce2f2586017859fb07
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 10:46:21 2026 -0500

    IMSnowBike

commit 4abb0a613ff9b643999f65259f37783aaac111d2
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 10:45:17 2026 -0500

    BWlog

commit b0d234f5b4d41b4f90c0f52b52660f743ccc216b
Author: Ryan Myers <ryan.p.myers@gmail.com>
Date:   Thu Feb 26 10:43:46 2026 -0500

    SPshop warnings

+++ src/newshadows.c
@@ -866,7 +866,7 @@ s32 shadows_draw2(Vtx* arg0, Gfx* dl, ObjectShadow* shadow, Object* obj, s32 arg
     func_80040FF8();
     tex_render_save_state();
     // @fake
-    if (&dl) {}
+    if ((s32)&dl) {}
     gSPGeometryMode(dl, 0xFFFFFF, G_FOG| G_CULL_BACK | G_SHADE | G_ZBUFFER);
     dl_apply_geometry_mode(&dl);
     if (sp1D4 != 0) {
